Revert "changed the numPars int for a float, in order to use a ref to it with ofxUI"
This reverts commit 9eae26f789427280c1df7802fa3eb9f54bafb14c.

+++ src/ofxParticleEmitter.h
@@ -44,7 +44,7 @@ class ofxParticleEmitter
         float life;
         float lifeSpread;
         
-        float numPars;
+        int numPars;
         
         ofColor color;
         ofColor colorSpread;
